# agent-team-workbench/contracts/web/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Team Workbench Web API
  version: 0.1.0
  description: |
    浏览器 ↔ Go Control Plane 协议（协议文档 §5）。
    约定：/api/v1 前缀；资源 ID 带类型前缀（ws_/agent_/wi_/run_/approval_/artifact_）；
    JSON 字段 snake_case；写命令要求 Idempotency-Key；可变资源带 version/ETag；
    错误一律 application/problem+json。
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /workspaces/{workspace_id}/work-items:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [work-items]
      operationId: listWorkItems
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: priority, in: query, schema: { type: string } }
        - { name: assignee, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemList" } } } }
    post:
      tags: [work-items]
      operationId: createWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemCreate" } } } }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
...
  /work-items/{work_item_id}/commands/assign:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: assignWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_profile_id, expected_version]
              properties:
                agent_profile_id: { type: string }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /work-items/{work_item_id}/commands/move:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: moveWorkItem
      description: 看板移动、排序与状态机校验。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, expected_version]
              properties:
                status: { $ref: "#/components/schemas/WorkItemStatus" }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /work-items/{work_item_id}/commands/block:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: blockWorkItem
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, message, expected_version]
              properties:
                code: { type: string }
                message: { type: string }
                source: { type: string }
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
  /work-items/{work_item_id}/commands/unblock:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: unblockWorkItem
      description: 解除阻塞回到 in_progress；恢复执行由用户显式创建新 Run。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version]
              properties:
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
  /work-items/{work_item_id}/runs:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    get:
      tags: [runs]
      operationId: listWorkItemRuns
      summary: 一个任务的全部 Run（对话轮次历史；不可覆盖的执行尝试）
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionRunList" } } } }
    post:
      tags: [runs]
      operationId: createRun
      description: 创建 ExecutionRun；立即 202，真正开始执行由 SSE 事件确认。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }, { $ref: "#/components/parameters/IfMatch" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RunCreate" } } } }
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/RunAccepted" } } } }
...
  /runs/{run_id}/commands/retry:
    parameters: [{ $ref: "#/components/parameters/RunId" }]
    post:
      tags: [runs]
      operationId: retryRun
      description: 总是创建新 Run（retry_of=当前 run），不覆盖终态；返回完整 ExecutionRun。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "202": { content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionRun" } } } }
...
  /runs/{run_id}/approvals/{approval_id}/commands/resolve:
    parameters: [{ $ref: "#/components/parameters/RunId" }, { $ref: "#/components/parameters/ApprovalId" }]
    post:
      tags: [approvals]
      operationId: resolveApproval
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                reason: { type: string }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Approval" } } } }
...
  /agent-profiles/{agent_profile_id}/task-sessions/reset:
    parameters: [{ $ref: "#/components/parameters/AgentProfileId" }]
    post:
      tags: [agent-profiles]
      operationId: resetTaskSession
      summary: 重置会话锚点（写墓碑；下一轮 Run 开新会话且不会被播种兜底复活）
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_key, adapter_id]
              properties:
                task_key: { type: string, description: 通常是 work item ID }
                adapter_id: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [reset] }
        "404": { $ref: "#/components/responses/Conflict" }
  /work-items/{work_item_id}/commands/accept:
    parameters: [{ $ref: "#/components/parameters/WorkItemId" }]
    post:
      tags: [work-items]
      operationId: acceptWorkItem
      description: 验收工作项（acceptance 决议；漂移检测后接受当前状态）。
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version]
              properties:
                expected_version: { type: integer }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    WorkspaceId: { name: workspace_id, in: path, required: true, schema: { type: string } }
    AgentProfileId: { name: agent_profile_id, in: path, required: true, schema: { type: string } }
    WorkItemId: { name: work_item_id, in: path, required: true, schema: { type: string } }
    RunId: { name: run_id, in: path, required: true, schema: { type: string } }
    ApprovalId: { name: approval_id, in: path, required: true, schema: { type: string } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: |
        写命令幂等键。同 key 同请求体：已完成则重放原响应（Idempotent-Replayed: true），
        执行中返回 409 idempotency_in_progress（retryable，稍后重试）；
        同 key 不同请求体返回 409 idempotency_conflict；5xx 后占位释放，可同 key 重试。
...
    IfMatch: { name: If-Match, in: header, required: false, schema: { type: string } }
    Cursor: { name: cursor, in: query, required: false, schema: { type: string } }
  responses:
    Conflict:
      description: 幂等冲突或乐观锁版本冲突
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unprocessable:
      description: 领域校验失败（状态机非法迁移等）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        retryable: { type: boolean }
        current_version: { type: integer }
...
    RuntimePreference:
      type: object
      properties:
        preferred: { type: string }
        fallbacks: { type: array, items: { type: string } }
        mode: { type: string, enum: [default, plan] }
        agent_preset: { type: string }
...
    WorkItemList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/WorkItem" } }
        next_cursor: { type: string, nullable: true }
    WorkItemCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/WorkItemStatus", default: todo }
        priority: { $ref: "#/components/schemas/Priority", default: medium }
        due_date: { type: string, format: date }
        agent_profile_id: { type: string, nullable: true }
...
    WorkItemStatus:
      type: string
      enum: [todo, in_progress, blocked, completed, cancelled]
    Priority:
      type: string
      enum: [low, medium, high, urgent]
    WorkItem:
      type: object
      properties:
        id: { type: string }
        workspace_id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/WorkItemStatus" }
        phase: { type: string, enum: [execution, review, acceptance], nullable: true }
        priority: { $ref: "#/components/schemas/Priority" }
        due_date: { type: string, format: date, nullable: true }
        agent_profile_id: { type: string, nullable: true }
        assignee_summary:
          type: object
          nullable: true
          properties:
            name: { type: string }
            avatar: { type: string }
        blocker:
          type: object
          nullable: true
          properties:
            code: { type: string }
            message: { type: string }
            source: { type: string }
            created_at: { type: string, format: date-time }
        runs_count: { type: integer }
        latest_run_id: { type: string, nullable: true }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RunCreate:
      type: object
      required: [agent_profile_id, input]
      properties:
        agent_profile_id: { type: string }
        runtime_preference: { $ref: "#/components/schemas/RuntimePreference" }
        requirements:
          type: object
          additionalProperties: { type: string, enum: [required, preferred, optional] }
        input:
          type: object
          required: [instruction]
          properties:
            instruction: { type: string }
            acceptance_criteria: { type: array, items: { type: string } }
    RunAccepted:
      type: object
      properties:
        run_id: { type: string }
        work_item_id: { type: string }
        status: { type: string }
        version: { type: integer }
        capability_snapshot_id: { type: string, nullable: true }
    ExecutionRun:
      type: object
      properties:
        id: { type: string }
        work_item_id: { type: string }
        agent_profile_id: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        provider: { type: string, nullable: true }
        runtime_label: { type: string, nullable: true }
        progress: { type: number, nullable: true }
        retry_of: { type: string, nullable: true }
        failure:
          type: object
          nullable: true
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RunStatus:
      type: string
      enum: [queued, starting, running, waiting_approval, interrupting, cancelling, reconnecting, succeeding, succeeded, interrupted, cancelled, lost, failed]
...
    Approval:
      type: object
      description: 以实现为准（httpapi approvalDTO）；不输出 requested_by/expires_at。
      properties:
        id: { type: string }
        run_id: { type: string }
        work_item_id: { type: string }
        kind: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [pending, approved, rejected, expired] }
        summary: { type: string }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolved_by: { type: string, nullable: true }
...
    ExecutionRunList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ExecutionRun" } }
...